Notes February 28th @ 2:01 AM
This section took me a while to work on. It took about 3 days to fully
understand what I was really learning.

1. First thing to do when starting this deleteItem function is setting
up the event listener for this action. Since this function is going to
be called on new items that have not yet been created I needed to create
an event delegator on the container class to prevent event bubbling
later on.

2. In the Global Controller I had to stage everything for this delete
function to work.

a. I created a variable called itemID that will store
the id that contains either "inc" or "exp". In to grab this id I had to
traverse the DOM using the parentNode method four times because that
is where that id will be stored.

b. Then I created a variable called splitID which removes the "-" and
creates a new array which will be the essentially be the "type" and
the "id" I am trying to capture

c. I created a variable called type and ID and stored their values
using the splitID[0]; and parseInt(splitID[1]); respectively.

3. After that I called the budgetCtrl.deleteItem(type, ID) method.

4. In the Budget Controller I created a deleteItem(type, ID) method
that will read the parameters type and ID from the Global Controller.

a. The deleteItem method has two variables called ids and index.

b. For the ids variable I utilized a new method called from the Array
prototype called Array.map(); What it does is that it reads the data
from the current data structure and it runs a mandatory function that
will return the current items in the array into a new array. This new
array is stored in the ids variable and it does not interfere with the
current data structure.

c. Now that the new array ids is created I can use it to store an
index number in the variable called index that I created that I will
be able to use when I want to delete that item from the data structure.

d. Finally created an if statement that will only run if there is an
index. Inside this if statement I will be deleting the item using the
splice method and using the positioning from the index variable.